% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data_inputs.R
\name{assert_dataframe}
\alias{assert_dataframe}
\title{Assert that input is a data frame}
\usage{
assert_dataframe(df)
}
\arguments{
\item{df}{An object to validate}
}
\value{
Stops execution if \code{df} is not a data frame.
}
\description{
Assert that input is a data frame
}
