% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data_inputs.R
\name{eval_factor_levels}
\alias{eval_factor_levels}
\title{Validate categorical factor levels in train/test labels}
\usage{
eval_factor_levels(train_labels, test_labels, min_count = 5)
}
\arguments{
\item{train_labels}{Factor vector of training labels}

\item{test_labels}{Factor vector of testing labels}

\item{min_count}{Minimum number of instances required per class (default: 5)}
}
\value{
Stops or warns if levels are mismatched or underrepresented
}
\description{
Validate categorical factor levels in train/test labels
}
