% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_target_plot.R
\name{make_target_plot}
\alias{make_target_plot}
\title{Create a bar plot for the target variable distribution}
\usage{
make_target_plot(df)
}
\arguments{
\item{df}{A data frame containing the target column "default_payment_next_month"}
}
\value{
A ggplot object showing the bar plot
}
\description{
Create a bar plot for the target variable distribution
}
