% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_abstraction.R
\name{prepare_knn_data}
\alias{prepare_knn_data}
\title{Prepare normalized predictors and extract labels}
\usage{
prepare_knn_data(df_train, df_test, target = "default_payment_next_month")
}
\arguments{
\item{df_train}{Training data}

\item{df_test}{Testing data}

\item{target}{Column name of target variable}
}
\value{
A list with normalized train/test data and labels
}
\description{
Prepare normalized predictors and extract labels
}
