% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R, R/modelling_abstraction.R
\name{split_data}
\alias{split_data}
\title{Split dataset into training and test sets}
\usage{
split_data(df, train_size = 0.8)

split_data(df, train_size = 0.8)
}
\arguments{
\item{df}{A data frame}

\item{train_size}{Proportion of data to use for training (default: 0.8)}
}
\value{
A list with train and test data frames

A list with 'train' and 'test' data frames
}
\description{
Split dataset into training and test sets

Split dataset into 80\% training and 20\% testing
}
