% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_abstraction.R
\name{split_dataset}
\alias{split_dataset}
\title{Split dataset into 80\% training and 20\% testing}
\usage{
split_dataset(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A list with 'train' and 'test' data frames
}
\description{
Split dataset into 80\% training and 20\% testing
}
