% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data_inputs.R
\name{warn_on_duplicates}
\alias{warn_on_duplicates}
\title{Warn if there are duplicate rows in either train or test}
\usage{
warn_on_duplicates(train, test)
}
\arguments{
\item{train}{A data frame containing training data}

\item{test}{A data frame containing test data}
}
\value{
Prints warning if duplicate rows are found
}
\description{
Warn if there are duplicate rows in either train or test
}
